Created DAO Refactoring entities

+++ petclinic-db/src/main/java/com/bas/petclinic/model/UserRole.java
@@ -11,16 +11,16 @@ public class UserRole {
 
     @Id
     @Column(name = "id")
-    private Byte id;
+    private Integer id;
 
     @Column(name = "name", length = 20)
     private String name;
 
-    public Byte getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Byte id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -38,7 +38,6 @@ public class UserRole {
         if (o == null || getClass() != o.getClass()) return false;
         UserRole userRole = (UserRole) o;
         return id.equals(userRole.id);
-
     }
 
     @Override
